requirements.txt
pandas
requests
llama-index-llms-ollama
fpdf
gradio
faiss-cpu
sentence-transformers